move all internal to components
