+++ components/qtmaterialcheckbox.cpp
@@ -1,5 +1,5 @@
 #include "qtmaterialcheckbox.h"
-#include "lib/qtmaterialcheckable_internal.h"
+#include "materiallib/qtmaterialcheckable_internal.h"
 #include "qtmaterialcheckbox_p.h"
 #include <QPropertyAnimation>
 #include <QSignalTransition>
