+++ components/qtmaterialradiobutton.cpp
@@ -1,5 +1,5 @@
 #include "qtmaterialradiobutton.h"
-#include "lib/qtmaterialcheckable_internal.h"
+#include "materiallib/qtmaterialcheckable_internal.h"
 #include "qtmaterialradiobutton_p.h"
 #include <QColor>
 #include <QPropertyAnimation>
